Refactor Main; utility functions are now in Utils
